ai-engine/requirements.txt
langchain-mistralai
python-dotenv
python-dateutil
pinecone
pypdf
fastapi
uvicorn
langgraph
langchain
psycopg2-binary
mcp
nest_asyncio
ddgs
dspy
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
gTTS
neo4j
